✨ feat: create new content page
